Moved reducer implementations to Core package
